+++ cpp/include/Ice/ios/iAPConnectionInfo.h
